refactor(apps-proxy): AJDA-3219 name the background-poll package honestly
Two problems with frameworkpoll, both mine.

Its docblock said the package existed because "their packages already
depend on each other in the other direction". That is not true: upstream
imports sessions, sessions imports nothing from upstream. There is no
cycle. The real constraint is narrower — the helper cannot live in
upstream, because sessions would then have to import it and *that* would
be the cycle — and I wrote it backwards.

And "framework" named an abstraction that does not exist: the contents
are two /_stcore/ paths, which is Streamlit and only Streamlit. A bare
Is() then left the package name carrying all the meaning and carrying it
badly.

Now dataapps/streamlit with IsBackgroundPoll, so both call sites read as
a sentence: streamlit.IsBackgroundPoll(req.URL.Path). Generalise the day
a second framework turns up. The docblock now gives the honest reason for
a separate package — it could live in sessions, but an auto-suspend rule
reading out of a package called sessions is an odd place to look.

Raised in review by @pepamartinec, who checked every import in the
package before saying the stated reason did not hold.
